Simplify and rename PlaceTest.kt
