--- src/test/kotlin/graphql/kickstart/tools/place/Entity.kt
@@ -1,3 +1,0 @@
-package graphql.kickstart.tools.place
-
-abstract class Entity(val id: String? = null)
