--- src/test/kotlin/graphql/kickstart/tools/place/OtherPlace.kt
@@ -1,7 +1,0 @@
-package graphql.kickstart.tools.place
-
-abstract class OtherPlace<R : Review<*>>(id: String? = null) : Place<R>(id) {
-
-    val other: String? = null
-
-}
