--- src/test/kotlin/graphql/kickstart/tools/place/Place.kt
@@ -1,8 +1,0 @@
-package graphql.kickstart.tools.place
-
-abstract class Place<R : Review<*>>(id: String? = null) : Entity(id) {
-
-    val name: String? = null
-    val reviews: MutableSet<R>? = null
-
-}
