--- src/test/kotlin/graphql/kickstart/tools/place/Place1.kt
@@ -1,3 +1,0 @@
-package graphql.kickstart.tools.place
-
-class Place1(id: String? = null) : OtherPlace<Review1>(id)
