--- src/test/kotlin/graphql/kickstart/tools/place/Place2.kt
@@ -1,3 +1,0 @@
-package graphql.kickstart.tools.place
-
-class Place2(id: String? = null) : OtherPlace<Review2>(id)
