--- src/test/kotlin/graphql/kickstart/tools/place/Review1.kt
@@ -1,3 +1,0 @@
-package graphql.kickstart.tools.place
-
-class Review1(id: String? = null) : Review<Place1>(id)
