--- src/test/kotlin/graphql/kickstart/tools/place/Review2.kt
@@ -1,3 +1,0 @@
-package graphql.kickstart.tools.place
-
-class Review2(id: String? = null) : Review<Place2>(id)
